consolidate StrIntEnum, use .title consistently
String representations now consistenly use "Title Strings" (each
word first letter is capitalized). This allows for direct mapping to
values that look nice in TK GUI and web UI.
